+++ backend/src/main/java/com/nuriwoolim/pagebackend/domain/user/controller/AuthController.java
@@ -8,11 +8,14 @@ import com.nuriwoolim.pagebackend.domain.user.dto.UserResponse;
 import com.nuriwoolim.pagebackend.domain.user.dto.UserSignupRequest;
 import com.nuriwoolim.pagebackend.domain.user.dto.VerificationResendResponse;
 import com.nuriwoolim.pagebackend.domain.user.service.AuthService;
+import com.nuriwoolim.pagebackend.domain.user.service.EmailVerificationService;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.Email;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -24,9 +27,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/auth")
+@Validated
 public class AuthController {
 
     private final AuthService authService;
+    private final EmailVerificationService emailVerificationService;
 
     @PostMapping("/signup")
     public ResponseEntity<Void> signup(
@@ -35,24 +40,10 @@ public class AuthController {
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
-    @GetMapping("/verify-email")
-    public ResponseEntity<Void> verifyEmail(@RequestParam String email, @RequestParam String code) {
-        authService.verifyEmail(email, code);
-
+    @GetMapping("/check-email")
+    public ResponseEntity<Void> checkEmail(@RequestParam @Email String email) {
+        authService.checkEmail(email);
         return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    @GetMapping("/send-verification")
-    public ResponseEntity<VerificationResendResponse> sendVerification(@RequestParam String email) {
-        VerificationResendResponse response = authService.sendVerificationEmail(email);
-        return ResponseEntity.ok(response);
-    }
-
-    @GetMapping("/resend-verification")
-    public ResponseEntity<VerificationResendResponse> resendVerification(
-        @RequestParam String resendToken) {
-        VerificationResendResponse response = authService.resendVerificationEmail(resendToken);
-        return ResponseEntity.ok(response);
     }
 
     @PostMapping("/login")
@@ -87,6 +78,28 @@ public class AuthController {
 
         return ResponseEntity.ok().build();
     }
+
+    @GetMapping("/verify-email")
+    public ResponseEntity<Void> verifyEmail(@RequestParam String email, @RequestParam String code) {
+        emailVerificationService.verifyEmail(email, code);
+
+        return ResponseEntity.status(HttpStatus.OK).build();
+    }
+
+    @GetMapping("/send-verification")
+    public ResponseEntity<VerificationResendResponse> sendVerification(@RequestParam String email) {
+        VerificationResendResponse response = emailVerificationService.sendVerificationEmail(email);
+        return ResponseEntity.ok(response);
+    }
+
+    @GetMapping("/resend-verification")
+    public ResponseEntity<VerificationResendResponse> resendVerification(
+        @RequestParam String resendToken) {
+        VerificationResendResponse response = emailVerificationService.resendVerificationEmail(
+            resendToken);
+        return ResponseEntity.ok(response);
+    }
+
     //TODO: 1. 비밀번호 찾기 메일 전송, 2. 비밀번호 찾기 메일 재전송, 3. 비밀번호 재설정
 
 }
